refactor: calcTimeDifference 유틸함수 작성
